LSPS1: Made JSON-example sensible
In the updated example the LSP is not supporting on-chain payments.
We can see this because `min_onchain_payment_size_sat` and
`min_onchain_payment_confirmations` are both set to `null`.

In the previous example it was indicated that we did not support
on-chain payments because but we still configured the `minimum_onchain_payment_confirmations`
variable