Add `temporary_channel_id` to `lsps2.buy` response
Previously, an LSPS2 client had no unambiguous way to link an incoming
channel opening from the LSP to a specific LSPS2 buy request it issued
earlier.

In order to give the client the capability to correlate which channel
was opened as a result of which request, we here require the LSP to send
the `temporary_channel_id` it is going to use for the channel opening in
the response to `lsps2.buy`.