LSPS1: Better semantics of `confirms_within_blocks`
The change is quite subtle but the intent becomes clear when you apply
it to a few examples.

## Before this commit

The `confirms_within_blocks` is an upper bound on the number of
- blocks between payment and funding-tx-confirmed (See [link](https://github.com/BitcoinAndLightningLayerSpecs/lsp/blob/d07cbb0bcdc1942c4970e4a2ac46f87c63dbdf33/LSPS1/README.md?plain=1#L449C7-L449C154))
- blocks between payment and channel_ready (See [link](https://github.com/BitcoinAndLightningLayerSpecs/lsp/blob/d07cbb0bcdc1942c4970e4a2ac46f87c63dbdf33/LSPS1/README.md?plain=1#L157C36-L157C119))

The `min_channel_confirmations` is the number of confirmations of the funding-tx before sending a `channel_ready`.

```
payment_confirmed            funding tx confirmed            channel_ready
    |                             |                               |
    |                             |                               |
    |<--confirms_within_blocks--->|                               |
    |                             |                               |
    |<--------------------confirms_within_blocks----------------->|
    |                             |                               |
    |                             |<--min_channel_confirmations-->|
    |                             |                               |
```

First, I want to point out that `min_channel_confirmations` is an LSP-wide configuration.
It should be specified in the order just like `confirms_within_blocks`.

The second problem is that the semantics `confirms_within_blocks` confusing.
It has two meanings that are spread out across the document.

The set of chosen parameters also poorly communicates the intent of a
transaction.

A client wants a 0-conf channel. Should it set `confirms_within_blocks=0`?
No, in that case the server MUST confirm the funding transaction in 0
blocks. This wouldn't be feasible.

Should the client set `confirms_within_blocks` sufficiently high to give the LSP-server
the opportunity to get the funding tx confirmed?

Maybe, but how would a client that wants to wait a few blocks and is
fine with batching communicate their intent. It can only set
`confirms_within_blocks` sufficiently high.

| Intent | `confirms_within_blocks` |
|--------|--------------------------|
| Client requests a zero-conf channel | 6 |
| Client wants to use a channel within 6 blocks and is prepared to wait for batching | 6 |

## Changes in this commit

We define `funding_confirms_within_blocks` as the amount of blocks between the
payment is confirmed and the funding tx is confirmed. (Single meaning)

We keep the definition of `min_channel_confirmations`. However, we require
the client to put this number explicitly in the order.

If `min_channel_confirmations=0` the server MUST open the channel immediately
after receiving payment

```
payment                 funding tx confirmed                        channel ready
  |                                  |                                    |
  |                                  |                                    |
  |<--funding_confirms_within_blocks>|                                    |
  |                                  |                                    |
  |                                  |<--required_channel_confirmations-->|
  |                                  |                                    |
  |                                  |                                    |
```

| `confirms_within_blocks` | `min_channel_confirmations` | Meaning                                                                                                       |
|---|---|----------------------------------------------------------------------------------------------------------------------------------------------------------------|
| 0 | 0 | Not allowed. The LSP cannot confirm a transaction within 0 blocks.                                                                                             |
| 6 | 0 | A zero-conf channel. The LSP should ensure the funding tx is confirmed within 6 blocks                                                                         |
| 6 | 1 | The LSP should ensure the funding tx is confirmed within 6 blocks after receiving the payment. It can still wait one more block until it sends `channel_ready` |