Move `{min,max}_payment_size_msat` fields into `opening_fee_params`
Previously, the `min_payment_size_msat` and `max_payment_size_msat`
fields were part of the `lsps2.get_info` response, but were not part of
the opening fee parameters the LSP commits to. It was however remarked
that the LSP *may* include them in the `promise` field.

Here, we make the latter non-optional and move the fields into the
`opening_fee_params`. This has three main benefits:

1) The payment size limits are non-negotiable, possibly per-request boundaries the LSP enforces and
   hence they should be clearly communicated and committed to via the `promise`.
2) The LSP may want to take different fee rates depending on the
   payment size limits, e.g., take a higher fee in exchange for allowing
   higher payment sizes.
3) *If* the LSP so far included them in the `promise` calculation, it
   was required to keep track of them for each `get_info` response it
   previously handed out to the users. Moving the fields into the
   `opening_fee_params` allows it to not keep any state between
   `lsps2.get_info` and `lsps2.buy`, decoupling these two steps further.