LSPS0: Drop 'Notes on Implementation', move common schemas reference up (#97)
* LSPS0: Drop 'Notes on Implementation'

As we're getting closer to finalizing the spec, we want to avoid
constantly having to touch it again.

As previously discussed implementation notes are quickly outdated (they
are in fact already, IIUC). We should therefore rather keep them in the
Wiki section of Github or similar, not in the spec document.

* Move Common Schemas reference up

As it's just a real brief paragraph, we might want to move the reference
to the commmon schemas doc up so readers are aware of it from the
beginning.