lsps1: Use int16 instead of uint8 (#113)
The variable `funding_confirms_withing_blocks` was of type `uint8`. This
means we allowed values from 0<=x<256. This corresponds to a duration of
42 hours.

We are limiting reasonable values and should increase the range of
allowed values.

Relates to https://github.com/BitcoinAndLightningLayerSpecs/lsp/issues/112
Relates to https://github.com/alexbosworth/balanceofsatoshis/issues/575